requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.6

Pillow>=8.4
requests>=2.29.0
transformers>=4.28.1
